Basic HTML page layout
- universal header across top to link back to lsst.org
- metadata in a left column. Include title/handle components based on
  the https://github.com/lsst-sqre/ltd-dasher project's dashboards.
- PDFObject in the right column

I'm using flexbox to layout (in row mode) for the metadata and pdf
columns. This is far easier than managing floats, particularly since the
PDF expands to fit the container, and that was interfering with the
floats.

NOTE: This is still a proof-of-concept. Test metadata is still being
shimmed into the config.