Add build date to the document info header
- Add the datetime of the PDF build.
- Use the calendar octicon.
- Integrate it with the version display/switcher through a <ul> below
  the header.

Adds pytz as a dependency in order to get a proper UTC date.

Now that we're just using the current time of the build. This works in a
CI continuous delivery mode, but doesn't work for old docs being
deployed now. Ultimately we should get the date from the last git commit
of fits in the tex document tree.