Fix data packaging to be pip installable
The rules are that all data needed by lander at runtime need to be
inside lander's package directory in order to be installed. Thus
templates/ is now inside lander/. Similarly, assets/ is changed to be
inside lander/ by changing the gulp assets build directories.

Also dropped the lander --license command since it's not easy to copy
the license into the lander/ package directory.