Use LsstLatexDoc's revision_datetime attribute
This attribute contributes the revision datetime based on the \date
command in the tex document, falling back to the Git timestamp and
finally 'now.'

This is an improvement over the previous behavior of always using 'now'
as the revision date. It provides flexibility to rebuild a document for
maintenance reasons without having the datetime drift.