Create metadata.jsonld file
This file is written adjacent to index.html as metadata.jsonld and
contains the full metadata about the document as schema.org/codemeta. It
includes the full text to enable back-end search. Because the full text
is present, I'm not including this document in the HTML for performance.

Replace metasrc with lsst-projectmeta-kit==0.3.0b0 to take advantage of
the new jsonld APIs.

Replace metasrc imports in the codebase with the new lsstprojectmeta
package.