Add LDM-151 integration test
This type of test exercise the end-to-end build of a document and
creation of a landing page. It simulates what we do on Travis, giving us
confidence that new lander deployments will work.

In this system, integration scripts are in the integration-tests
directory. They can be run from the main make file.