Skip if expected variables aren't set for travis
This is likely the best way at this point to determine if the build is
happening from a fork on Travis. Note that the TRAVIS_SECURE_ENV_VARS
environment variable doesn't help here because it's still true even once
those variables have been decrypted.

This code really needs to be refactored so we aren't running sys.exit
from deep within the config code.