Create DocumentMetadata pydantic model
This provides a consistent data model for document parsing plugins to
adhere to.

The EncodedString model exists to allow the abstract to include an HTML
version with links and other HTML inline syntax in place. The HTML is
sanitized with the bleach package so that unnecessary/unsafe tags are
removed automatically.