Resolve pandoc and panflute dependency matrix
panflute has fairly tight pandoc dependency requirements, see

https://github.com/sergiocorreia/panflute#supported-pandoc-versions

Since we're trying to use the py-pandoc package to get pandoc with
version constraints, we're limited to pandoc 2.9 because py-pandoc does
not have a pandoc 2.11 release.

In turn, this means we need to constrain what version of panflute is
used, to the 1.2 series.

Finally, in tox we block any external pandoc that might be on the
developer's machine to ensure that the pandoc supplied with py-pandoc is
used.