Introduce MVP for template plugins
Template plugins are responsible for rendering the HTML for the landing
page site given the PDF, metadata, and settings.

This is an MVP featuring plugin discovery through the lander.template
setuptools entrypoint, a baseclass for template plugins, a basic
rendering pipeline for the site that copies files and renders Jinja
templates.

To test the plugin API, this commit also adds a "minimalist" template
that might serve as the built-in default template.