Add lander.js app and bundling pipeline
lander.js contains the application code for running PDFObject to embed a
PDF in the landing page. The improvement from the previous generation of
the code is that a ?page=N search string can be set in the URL to
permalink to a specific page of the PDF.

The gulpfile has been repurposed to package these specific assets within
individual themes. Each theme can now have its own webpack
configuration file that is used for specific JS bundling tasks within
gulp.