Add Contributor class, used for authors
Contributor is a subclass of Person that includes an additional "role"
attribute. This role can be used to annotate an author's specific role.
This can be useful in some author lists.

The authors attribute of the DocumentMetadata listing now uses the
Contributor type.

The original Author class can still be used by plugin-specific
subclasses of DocumentMetadata that don't have a need for role metadata
and just need to identify a person.