Pre-install pandoc in GitHub Actions with brew
It looks like pypandoc's download_pandoc function isn't working. We'll
continue to use the decorator to check that pandoc exists, but for now
recommend that users install pandoc separately.

apt-get install was giving pandoc 2.5, which is too old for panflute.
homebrew, however, gives us up-to-date versions of pandoc, which is what
we need.