Fix setting date metadata from the command line
The dictionary key that date was being set too for cli_configs didn't
match the attribute in the Configuration dictionary, and therefore
wasn't used in the configuration.

Because build_datetime is a datetime, if a user provides a date as
YYYY-MM-DD through the command line, the new pre-validator converts that
date into a datetime instance.